# ENH: Improve performance for df.__setitem__ with list-like indexers