--- a/pandas/core/indexing.py
+++ b/pandas/core/indexing.py
@@ -663,17 +663,10 @@ class _LocationIndexer(NDFrameIndexerBase):
             and not com.is_bool_indexer(key)
             and all(is_hashable(k) for k in key)
         ):
-            for i, k in enumerate(key):
-                if k not in self.obj:
-                    if value is None:
-                        self.obj[k] = np.nan
-                    elif is_array_like(value) and value.ndim == 2:
-                        # GH#37964 have to select columnwise in case of array
-                        self.obj[k] = value[:, i]
-                    elif is_list_like(value):
-                        self.obj[k] = value[i]
-                    else:
-                        self.obj[k] = value
+            keys = self.obj.columns.tolist()
+            keys.extend([k for k in key if k not in self.obj])
+            data = self.obj._mgr.reindex_axis(keys, 0)
+            object.__setattr__(self.obj, "_mgr", data)
 
     def __setitem__(self, key, value):
         if isinstance(key, tuple):