@jbrockmendel 2020-11-30
@phofl did you determine that this isn't making a copy?