@phofl 2020-11-30
Not with a test, but I tested this in code.
```
x = self.obj
self.obj._mgr = self.obj._mgr.reindex_axis(keys, 0)
```

This also changes x, so that is not a copy, isn't it?