@jbrockmendel 2020-11-30
> This also changes x, so that is not a copy, isn't it?

Assuming homogeneous dtype for now, what im asking for is:

```
values = self.obj.values
self.obj._mgr = self.obj._mgr.reindex_axis(keys, 0)
new_values = self.obj.values

assert new_values is values   # <-- will raise if we made a copy
```