@jreback 2020-11-30
i am not sure this is possible w/o a copy