@phofl 2020-11-30
This does make a copy. Did not know, that we have to compare values instead of the object itself.