@phofl 2020-11-30
Actually this raised previously too, so at least not a regression