@jbrockmendel 2020-11-30
> Did not know, that we have to compare values instead of the object itself.

Note: the example above I assumed homogeneous dtypes.  More generally, we need to have `obj._mgr.blocks[n].values` be unchanged for each pre-existing n.

> i am not sure this is possible w/o a copy

A few things need to happen for it to be possible:

1) add `consolidate: bool = True` keyword to `BlockManager.reindex_axis` that then passes consolidate=consolidate to `reindex_indexer`
2) add `only_slice: bool = False` to `reindex_axis` and `reindex_indexer` so it can be passed to `_slice_take_blocks_ax0`
3) the `reindex_axis` call in this PR passes `consolidate=False, only_slice=True`