@simonjayhawkins 2020-11-30
@meeseeksdev backport 1.1.x