@simonjayhawkins 2020-11-30
https://github.com/pandas-dev/pandas/pull/38181#issuecomment-735682557

> @phofl #38148 (comment) was this necessary? do we know what commit caused the performance regression. the existing code on 1.1.x looks much simpler to what was replaced on master.

If we don't backport this, we will need to move the release note on master