@jorisvandenbossche 2020-11-30
As @jbrockmendel noted, this *can* change the copy/view semantics in certain cases. 

One example:

```python
df = pd.DataFrame({'a': [1, 2, 3], 'b': [4, 4, 6]})
# get one column as a view of df
s = df['a']
# add columns with list-like indexer
df[['c', 'd']] = np.array([[.1, .2], [.3, .4], [.4, .5]])
# edit in place the first column to check view semantics
df.iloc[0, 0] = 100
```

on master this gives:
```
In [8]: df
Out[8]: 
     a  b    c    d
0  100  4  0.1  0.2
1    2  4  0.3  0.4
2    3  6  0.4  0.5

In [9]: s
Out[9]: 
0    1
1    2
2    3
Name: a, dtype: int64
```

where "a" and "b" columns were now copied in the `df[['c', 'd']] = ...` operation (shown by `s` not being updated). 

While on pandas 1.1.4, this copy didn't happen, and the series was actually updated:

```
In [6]: df
Out[6]: 
     a  b    c    d
0  100  4  0.1  0.2
1    2  4  0.3  0.4
2    3  6  0.4  0.5

In [7]: s
Out[7]: 
0    100
1      2
2      3
Name: a, dtype: int64
```

This of course closely relates to the recent discussions about improving the copy/view semantics (as currently those semantics are both not clear and largely untested).

(now the above is certainly a specific example to trigger the issue. In many cases we would also do a copy, eg if there would already have been float columns present in the 
Given those ungoing discussions, we maybe shouldn't change those semantics now (I think the goal of of the discussions would certainly be