@jbrockmendel 2020-12-01
what if instead of calling reindex_axis we called reindex_indexer and passed `indexer=slice(len(self.obj.columns))`?